+++ multi-level-dropdowns/src/components/Menu/Menu.tsx
@@ -1,20 +1,21 @@
-import { FC, HTMLAttributes } from "react";
+import { FC } from "react";
 
 import SubMenu from "./SubMenu";
 import List from "../core/List";
 import ListItem from "../core/ListItem";
+import useMenu from "../../hooks/useMenu";
 
 import classes from "./Menu.module.css";
 
-interface MenuProps extends HTMLAttributes<HTMLElement> {
-  menu: Array<MultiLevelDropdown>;
-}
+const Menu: FC = () => {
+  const { menu } = useMenu();
 
-const Menu: FC<MenuProps> = ({ menu, ...rest }) => {
+  if (!menu) return null;
+
   return (
-    <nav className={classes.nav} {...rest}>
+    <nav className={classes.nav}>
       <List>
-        {menu.map(({ id, name, children }) => (
+        {menu?.map(({ id, name, children }) => (
           <ListItem key={id}>
             {name}
             <SubMenu menu={children} />
