--- multi-level-dropdowns/src/pages/Main/Main.module.css
@@ -1,7 +1,0 @@
-.main {
-  margin: 2.25rem 1.25rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2rem;
-}
